introduce property expression
